Add files via upload
This project aims to assist the Technical Program Manager from PT X in retrieving information by prioritizing on technical issues topic from the review data submitted by users to support his decision on refinement plan. Review data will be classified by Support Vector Machine (SVM) and Naïve Bayes (NB) algorithm which can predict the class of data whether it is technical or non-technical matters. After the review data on technical issues is predicted, topic modelling will be performed using Latent Dirichlet Allocation (LDA) algorithm. The output of the LDA model then visualized on the LDAvis package to determine the popularity level of the topic, called topic priority, which is based on the estimated proportion of the number of words on the topic over the total number of words in the entire document.